+++ src/main/java/com/backend/workflow/repository/TipoPedidoRepository.java
@@ -3,5 +3,5 @@ package com.backend.workflow.repository;
 import com.backend.workflow.entity.TipoPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoPedidoRepository extends JpaRepository<TipoPedido, Integer>{
+public interface TipoPedidoRepository extends JpaRepository<TipoPedido, Long>{
 }
